connectu-backend/connectuapi/requirements.txt
asgiref==3.5.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.16.0
channels==4.0.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.5
defusedxml==0.8.0rc2
Django==4.1
django-cors-headers==3.13.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djoser==2.2.0
h11==0.14.0
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
oauthlib==3.2.2
Pillow==9.2.0
psycopg2==2.9.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==23.3.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==23.1.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.2
twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing-extensions==4.8.0
tzdata==2022.2
urllib3==1.26.12
uvicorn==0.23.1
websockets==11.0.3
wget==3.2
zope.interface==6.1


anyio==3.7.1
exceptiongroup==1.1.2
httptools==0.6.0
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
uvloop==0.17.0
watchfiles==0.19.0